// services/projects/src/rfis/dto/update-rfi.dto.d.ts
import { RFIStatus, RFIPriority } from './create-rfi.dto';
export declare class UpdateRFIDto {
    title?: string;
    question?: string;
    answer?: string;
    assignedTo?: string;
    dueDate?: string;
    status?: RFIStatus;
    priority?: RFIPriority;
    metadata?: Record<string, any>;
}
//# sourceMappingURL=update-rfi.dto.d.ts.map